bezig met verwijderen user

+++ src/pages/profile/Profile.js
@@ -1,19 +1,22 @@
 import styles from './Profile.module.css'
 import React, {useContext, useEffect} from 'react';
 import ContentCard from "../../components/contentCard/ContentCard";
-// import profilePicture from '../../assets/profile_picture.jpeg'
 import editIcon from '../../assets/edit_task_icon.svg'
+import deleteAccountIcon from '../../assets/delete_account.svg'
 import {useHistory} from "react-router-dom";
 import Icon from "../../components/icon/Icon";
 import InnerOuterContainer from "../../components/innerOuterContainer/innerOuterContainer";
 import {AuthContext} from "../../context/AuthContext";
 import {useForm} from "react-hook-form";
+import specialtiesString from "../../helpers/specialtiesString";
 //Firebase imports
 import {getStorage, ref} from "firebase/storage";
-import {db, storage} from "../../Firebase";
+import {authFirebase, db, storage} from "../../Firebase";
 import {uploadBytes} from "firebase/storage";
 import {deleteObject, getDownloadURL} from "firebase/storage";
 import {deleteDoc, doc, updateDoc} from "firebase/firestore";
+import WarningPopup from "../../components/warningPopup/WarningPopup";
+import {deleteUser} from "firebase/auth";
 
 function Profile({setCurrentPage}) {
 
@@ -22,6 +25,7 @@ function Profile({setCurrentPage}) {
     const [error, setError] = React.useState({error: false, message: ""});
     const [loading, toggleLoading] = React.useState(false);
     const [profilePictureUrl, setProfilePictureUrl] = React.useState("");
+    const [warningPopup, toggleWarningPopup] = React.useState(false);
 
     const history = useHistory();
     const {register, reset, formState: {errors}, watch, control, handleSubmit} = useForm();
@@ -47,6 +51,7 @@ function Profile({setCurrentPage}) {
             const pictureReference = await getDownloadURL(ref(storage, user.profilePicture))
             setProfilePictureUrl(pictureReference)
         }
+
         fetchProfilePicture()
     }, [user.profilePicture])
 
@@ -75,8 +80,10 @@ function Profile({setCurrentPage}) {
 
             // Delete old profile picture
             const desertRef = ref(storage, user.profilePicture);
-            await deleteObject(desertRef);
-
+            // Don't delete the default profile picture
+            if (desertRef.name !== "default_profile_picture.jpeg") {
+                await deleteObject(desertRef);
+            }
             toggleAuth({
                 ...auth,
                 user: {
@@ -84,12 +91,36 @@ function Profile({setCurrentPage}) {
                     profilePicture: 'profilePictures/' + user.id + "_" + data.file[0].name,
                 },
             })
+
         } catch (e) {
             console.error(e);
             setError({error: true, message: "Het uploaden is niet gelukt. Probeer opnieuw"})
         }
         toggleLoading(false);
         toggleUploadCard(false);
+    }
+
+    async function handleDeleteAccountClick() {
+        try {
+            const user = authFirebase.currentUser;
+            await deleteUser(user);
+
+            // TODO: verwijder gegevens van de user: sessies, profielfoto
+        } catch (e) {
+            console.error(e);
+            setError({
+                error: true,
+                message: "Er is iets misgegaan met het verwijderen van het account. Probeer het opnieuw"
+            })
+        }
+
+
+        deleteUser(user).then(() => {
+            // User deleted.
+        }).catch((error) => {
+            // An error ocurred
+            // ...
+        });
     }
 
     return (
@@ -131,8 +162,9 @@ function Profile({setCurrentPage}) {
                     </div>
                     <div className={styles["bottom-section"]}>
                         <p className={styles["profile-details"]}>Email: <span>{user.email}</span></p>
-                        <p className={styles["profile-details"]}>Functie: <span>{user.function.charAt(0).toUpperCase() + user.function.slice(1)}</span></p>
-                        <p className={styles["profile-details"]}>Specialiteiten: <span>{user.specialties.join(", ")}</span>
+                        <p className={styles["profile-details"]}>Functie: <span>{user.function.charAt(0).toUpperCase() + user.function.slice(1)}</span>
+                        </p>
+                        <p className={styles["profile-details"]}>Specialiteiten: <span>{specialtiesString(user.specialties)}</span>
                         </p>
                     </div>
                 </section>
@@ -145,7 +177,23 @@ function Profile({setCurrentPage}) {
                             history.push("/profiel-wijzigen")
                         }}
                     />
+                    <Icon
+                        text="Account verwijderen"
+                        image={deleteAccountIcon}
+                        alt="delete-account"
+                        onClick={() => {
+                            toggleWarningPopup(true);
+                        }}
+                    />
                 </figure>
+                {error.error && <span className={styles.error}>{error.message}</span>}
+
+                {warningPopup &&
+                    <WarningPopup
+                        text="Deze actie zal uw account en bijbehorende geregistreerde uren definitief verwijderen."
+                        toggleWarningPopup={toggleWarningPopup}
+                        handleButtonClick={handleDeleteAccountClick}
+                    />}
             </ContentCard>
         </InnerOuterContainer>
     );
